Refactor final scripts - [ ] Rip of jQuery
Use Huge for
- [x] Scripts
- [x] Styles
- [ ] SEO

Fixes #231
