+++ assets/js/main.js
@@ -1,0 +1,27 @@
+console.log("have js")
+menu_items = document.querySelectorAll(".m-standard-nav li").forEach(item => {
+  if (window.location.href.indexOf(item.getAttribute('data-taxonomyname')) !== -1) {
+    item.classList.add('active')
+  }
+});
+document.querySelectorAll(".share-trigger").forEach(item => {
+  item.addEventListener("click", function() {
+    console.log("click")
+    console.log(item)
+    const sibling = item.nextElementSibling
+    sibling.classList.toggle('show')
+  })
+});
+
+/* Adding caption to images in news pages */
+var richTextImages = document.querySelectorAll(".user-content img");
+if (richTextImages) {
+    richTextImages.forEach(function(element) {
+    var title = element.getAttribute("title");
+    var caption =
+        "<figcaption>" + title + "</figcaption>";
+    if (title !== null) {
+        element.insertAdjacentHTML("afterend", caption);
+    }
+    });
+}
