+++ assets/jsconfig.json
@@ -3,6 +3,7 @@
   "baseUrl": ".",
   "paths": {
    "*": [
+    "*",
     "../../../../../../var/folders/t7/9pbqz9_17llf0v0x7k3nxnkm0000gn/T/hugo_cache/modules/filecache/modules/pkg/mod/github.com/the!new!dynamic/hugo-module-tnd-search@v0.2.4/assets/*"
    ]
   }
