+++ assets/sass/modules/_nav.scss
@@ -300,4 +300,9 @@ a.emergency{
     text-decoration: none;
 }
 
+@media (min-width: $medium) {
+    #main-nav[style]{
+        display: block !important;
+    }
+}
 [x-cloak] { display: none !important; }
