--- data/assets.json
@@ -1,18 +1,0 @@
-{
-  "app": {
-    "js": "/dist/app.34e53548e8a46f076851.js",
-    "jsIntegrity": "sha256-2ejVBkcmQyIsqH97ENzRVSUnjdvSGZS1Xa+4DNqREi0="
-  },
-  "instagram": {
-    "js": "/dist/instagram.8c4138beca5ec146033d.js",
-    "jsIntegrity": "sha256-I/NZMRpYKLC8CX1NIN9K2eDePENikvPyUsl6VrEt1QQ="
-  },
-  "main": {
-    "js": "/dist/main.e6e6cf4b17bfe733192a.js",
-    "jsIntegrity": "sha256-Vd1pmmU0cybR+dYZSawrKHmoHTVjVwShmuDn6EuMrP0="
-  },
-  "search": {
-    "js": "/dist/search.1b7e0898715861d62536.js",
-    "jsIntegrity": "sha256-Jol2Fxw6M+069+sEILQOgLWgnu0J8n43ERhWH/mCbTk="
-  }
-}
