+++ go.mod
@@ -3,9 +3,12 @@ module github.com/theNewDynamic/oregonsadventurecoast.com
 go 1.14
 
 require (
+	github.com/theNewDynamic/huge v0.1.5 // indirect
 	github.com/theNewDynamic/hugo-module-tnd-env v0.0.0-20210512181247-90dd1a287016 // indirect
 	github.com/theNewDynamic/hugo-module-tnd-func v0.0.0-20210316194241-a6d99cfae5a6 // indirect
 	github.com/theNewDynamic/hugo-module-tnd-icons v0.1.7 // indirect
+	github.com/theNewDynamic/hugo-module-tnd-imgix v0.2.2 // indirect
+	github.com/theNewDynamic/hugo-module-tnd-media v0.0.0-20220330175441-4aeb80c7e616 // indirect
 	github.com/theNewDynamic/hugo-module-tnd-search v0.2.4 // indirect
 	github.com/theNewDynamic/hugo-module-tnd-seo v0.1.10 // indirect
 	github.com/theNewDynamic/hugo-module-tnd-snipcart v0.0.0-20210108153238-b64a5aee28af // indirect
