+++ layouts/_default/home.html
@@ -314,8 +314,3 @@
 
 
 {{ end }}
-
-
-{{ define "footerJs" }}
-{{ partial "script" "vendors/cycle2.min.js" }}
-{{ end }}
