+++ layouts/partials/hd.html
@@ -29,7 +29,7 @@
 	<nav class="m-nav" x-data="{ openNav: false }">
 		<div class="mobile-menu clearfix">
 			<!-- <span class="search"><i class="fas fa-search"></i></span> -->
-			<span class="menu-toggle" @click="openNav = !openNav"><span class="menu-link g-menu-x"><i class="bar bar-top"></i><i class="bar bar-mid"></i><i class="bar bar-btm"></i></span> <span class="menu-text">MENU</span></span>
+			<span class="menu-toggle" @click="openNav = !openNav"><span class="menu-link g-menu-x" :class="openNav ? '-active' : ''"><i class="bar bar-top"></i><i class="bar bar-mid"></i><i class="bar bar-btm"></i></span> <span class="menu-text">MENU</span></span>
 		</div>
 		<div class="inner clearfix" id="main-nav" 
 		x-cloak
