--- resources/_gen/assets/scss/sass/main.scss_48b060fe05b0a273d182ef83c0605941.json
@@ -1,1 +1,0 @@
-{"Target":"sass/main.min.75f45eb63c1854f1e1f633b8e8f29586dfe5b202cce4c36cfd8bc01d849cb65c.css","MediaType":"text/css","Data":{"Integrity":"sha256-dfRetjwYVPHh9jO46PKVht/lsgLM5MNs/YvAHYSctlw="}}
