+++ resources/_gen/assets/scss/sass/main.scss_f300667da4f5b5f84e1a9e0702b2fdde.json
@@ -1,0 +1,1 @@
+{"Target":"sass/main.css","MediaType":"text/css","Data":{}}
