--- src/js/app.js
@@ -1,10 +1,0 @@
-// JS Goes here - ES6 supported
-if (window.netlifyIdentity) {
-  window.netlifyIdentity.on("init", user => {
-    if (!user) {
-      window.netlifyIdentity.on("login", () => {
-        document.location.href = "/admin/";
-      });
-    }
-  });
-}
