--- src/js/common/constants.js
@@ -1,15 +1,0 @@
-export const PAGINATION_DEFAULTS = {
-    page: 1,
-    show: 9,
-    total: 0
-};
-
-export const PAGINATION_ACTIONS = {
-    next: "next",
-    prev: "prev"
-};
-
-export const FILTER_OPTION = {
-    add: "add",
-    remove: "remove"
-};
