--- src/js/common/get-filter-match-type.js
@@ -1,7 +1,0 @@
-export default class GetFilterMatchType {
-    constructor() {}
-
-    getType(options, key) {
-        return options[key];
-    }
-}
