--- src/js/main.js
@@ -1,57 +1,0 @@
-(function($) {
-    // [x] Main nav toggle
-    $('.menu-togglex').click(function(){
-        $(this).find('.g-menu-x').toggleClass('-active');
-        $('#main-nav').slideToggle('fast');
-    });
-
-    // [x] Sidebar nav toggle for mobile
-    $('.m-standard-navX').find('h3').append('<span class="menu-link g-menu-x"><i class="bar bar-top"></i><i class="bar bar-mid"></i><i class="bar bar-btm"></i></span>');
-    $('.m-standard-navX').find('h3').click(function(){
-        $(this).find('.g-menu-x').toggleClass('-active');
-        $(this).next('ul').slideToggle();
-        $(this).next('.nav-group-display').slideToggle();
-    });
-    
-
-    // PDF & External Links in new window
-    $('#bd').find('a').each(function(){
-        let href = $(this).attr('href');
-        if (href !== undefined && href.includes('.pdf')) {
-            $(this).attr('target', '_blank');
-        } else if (
-                href !== undefined &&
-                !href.includes('oregonsadventurecoast.com') &&
-                !href.includes('localhost') &&
-                href.charAt(0) !== '/' &&
-                href.charAt(0) !== '#' &&
-                !href.includes('mailto:')) {
-            $(this).attr('target', '_blank');
-        }
-    });
-
-    // Sidebar nav highlight current page
-    $('.m-standard-nav').find("li").each(function(index, item){
-        if (window.location.href.indexOf($(this).data("taxonomyname")) !== -1) {
-            $(this).addClass('active');
-        }
-    });
-
-    // Share button trigger
-    $('.share-trigger').click(function(){
-        $(this).next('.m-share').toggle();
-    });
-})(jQuery);
-
-/* Adding caption to images in news pages */
-var richTextImages = document.querySelectorAll(".user-content img");
-if (richTextImages) {
-    richTextImages.forEach(function(element) {
-    var title = element.getAttribute("title");
-    var caption =
-        "<figcaption>" + title + "</figcaption>";
-    if (title !== null) {
-        element.insertAdjacentHTML("afterend", caption);
-    }
-    });
-}
